Rework Dockerfile to build in separate container
Currently, everything is built/assembled in the same docker
container, which results in image sizes being >2GB. This is
undesireable for such a minimal project.

This patch modifies the Dockerfile logic such that a separate
container is added for all build tasks, and as-such removes any
build-dependencies from the production container itself. This
shrinks the image size to <500MB, which is much better than
the current state.